Keyboard and Controller input both supported
Keyboard will only work with no controllers connected. This is all thanks to the deviceNumber of input events being 0 for both keyboard and controller, so there's no sane way to differentiate those 🫠

If a controller is connected mid game the player will control buggy. Not a big issue for me long term but will need to keep this in mind.